+++ instrumentation/tracing.go
@@ -174,4 +174,3 @@ func AddSecurityAttributes(span trace.Span, clientIP string) {
 		SetSpanAttributes(span, attribute.String(AttrClientIP, clientIP))
 	}
 }
-